% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\docType{package}
\name{BAStag-package}
\alias{BAStag-package}
\title{BAS tag data processing}
\description{
Basic data processing for British Antarctic Survey archival tags.
Provides facilities for importing and plotting data recorded by the
BAS tags, and in particular detecting times of twilight from the
recorded light data.
}
\author{
Simon Wotherspoon, Micheal Sumner and Simeon Lisovski.
}

