% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{crepuscularEdit}
\alias{crepuscularEdit}
\title{Edit crepuscular segments}
\usage{
crepuscularEdit(tagdata, twilights, offset = 0, extend = 6,
  threshold = NULL, lmax = 64, zlim = c(0, lmax), point.cex = 0.5,
  width = 12, height = 4, palette = defaultPalette[c(5, 2, 9, 3, 4, 1,
  1)])
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{twilights}{dataframe of twilight times as generated by
\code{\link{findCrepuscular}}.}

\item{offset}{the starting hour for the vertical axes.}

\item{extend}{the period (in hours) before and after twilight for
which the light profile should be plotted.}

\item{threshold}{the light threshold that defines twilight.}

\item{lmax}{the maximum light level to plot.}

\item{zlim}{the range of light levels to plot in images.}

\item{point.cex}{expansion factor for plot points.}

\item{width}{width of the interface windows.}

\item{height}{height of the interface windows.}

\item{palette}{a colour palette of 8 colours.}
}
\value{
the dataframe of edited twilights, with columns
\item{\code{Twilight}}{edited times of twilight}
\item{\code{Rise}}{logical indicating sunrise}
\item{\code{Start}}{date of first observation in the crepuscular segment}
\item{\code{End}}{date of last observation in the crepuscular segment}
}
\description{
Interactively edit crepuscular intervals
}
\details{
Interactively edit the crepuscular intervals based on the light
profile.  A plot of the estimated sunrise and sunset intervals is
displayed, and the user can select the twilight to be edited with
a left mouse click.

The light profile for the selected twilight is shown in a separate
window, and the selected segments of the light profile are
highlighted.  The corresponding light profiles from the preceeding
and following days are also shown for reference.

The user may select a new candidate interval by clicking and
dragging with the left mouse button.  Individual points may be
selected or deselected with the right mouse button.

Note however, no actual change to the selection is made until the
candidate edits are accepted depressing the 'a' key.

Each twilight may be marked with the an integer 0 to 9 with the
numeric keys. By default each day is given the mark 0.

In either window
\tabular{ll}{
'q' \tab Quits, returning the dataframe of edited twilight segments \cr
'a' \tab Accepts the candidate edit \cr
'r' \tab Resets the selection \cr
'i' \tab Toggles the display of the light image \cr
'p' \tab Toggles the display of individual points \cr
'l' \tab Toggles the display of surrounding profiles \cr
'+'/'-' \tab Zoom in or out \cr
'Left arrow' \tab Jump to previous twilight \cr
'Right arrow' \tab Jump to next twilight \cr
'0'-'9' \tab Mark this twilight \cr
}
}
\seealso{
\code{\link{preprocessLight}}
}
