% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAStag.R
\name{driftAdjust}
\alias{driftAdjust}
\title{Clock Drift Adjustment}
\usage{
driftAdjust(time, start, end)
}
\arguments{
\item{time}{a vector of POSIXct times.}

\item{start}{new start time as POSIXct.}

\item{end}{new end time as POSIXct.}
}
\description{
Adjust time for clock drift
}
\details{
Linearly rescale a sequence of dates to a new start and end time
to correct for clock drift in the tag.
}

