% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{edit_set_twilight}
\alias{edit_set_twilight}
\alias{edit_undo_set}
\alias{edit_set_marker}
\alias{edit_search_twilights}
\alias{edit_insert_twilight}
\alias{edit_undo_insert}
\alias{edit_toggle_deletion}
\alias{edit_set_interval}
\alias{edit_init_undo}
\title{Twilight Editing}
\usage{
edit_set_twilight(twilights, index, gmt)

edit_undo_set(twilights, index)

edit_set_marker(twilights, index, marker, fixed, zenith)

edit_search_twilights(tagdata, threshold, include, exclude, extend, dark.min)

edit_insert_twilight(twilights, gmt, rise)

edit_undo_insert(twilights)

edit_toggle_deletion(twilights, index)

edit_set_interval(twilights, minGMT, maxGMT)

edit_init_undo(twilights)
}
\arguments{
\item{twilights}{a dataframe of twilights}

\item{index}{the index of the twilight to modify.}

\item{gmt}{a string corresponding to a GMT time.}

\item{marker}{the integer marker.}

\item{fixed}{a two column array of fixed locations.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)}

\item{threshold}{the light threshold that defines twilight.}

\item{include}{a vector of times. Nights that span these
times are included in the search.}

\item{exclude}{a vector of times. Nights that span these
times are excluded from the search.}

\item{extend}{a time in minutes. The function seeks periods of
darkness that differ from one another by 24 hours plus or minus
this interval.}

\item{dark.min}{a time in minutes. Periods of darkness shorter
than this interval will be excluded.}

\item{rise}{is the twilight a sunrise.}

\item{minGMT}{a string corresponding to a GMT time.}

\item{maxGMT}{a string corresponding to a GMT time.}
}
\value{
The modified twilight dataframe.
}
\description{
These functions are used to edit the dataframe of twilight times.
}
\details{
These functions are used in the auditing script generated by
\code{\link{preprocessLight}}.
}
