% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{extractCrepuscular}
\alias{extractCrepuscular}
\title{Extract crepuscular light segments}
\usage{
extractCrepuscular(
  tagdata,
  twilights,
  adjust.interval = 0,
  filter.threshold = NULL
)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{twilights}{a dataframe generated by
\code{\link{findCrepuscular}}.}

\item{adjust.interval}{timing adjustment for sunset intervals.}

\item{filter.threshold}{\code{NULL} or a positive threshold.}
}
\value{
Returns a dataframe with columns
\item{\code{Date}}{the date and time of the observation}
\item{\code{Light}}{the recorded light level}
\item{\code{Segment}}{an integer identifying the segment}
}
\description{
Extract crepuscular light segments
}
\details{
Extract the crepuscular subsets of the light data, given a
dataframe of start and end points generated by
\code{\link{findCrepuscular}}.

The \code{adjust.interval} argument can be used to specify a
timing adjustment for tags that report the maximum light level
observed in the preceeding sampling interval. If this argument is
zero, no adjustment is made, otherwise the timestamps of sunset
intervals will be adjusted by this interval to compensate for the
recording behaviour of the tag.

If \code{filter.threshold} is non \code{NULL}, the light levels
are filtered to remove shaded observations.  The light levels at
sunrise and sunset are assumed to be monotonic increasing, and the
running maximum in the direction of day is calculated.  Light
levels that are \code{filter.threshold} units below the running
maximum are removed.
}
