% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{findChunk24}
\alias{findChunk24}
\title{Find extended periods of light or dark}
\usage{
findChunk24(tagdata, threshold)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{threshold}{the light threshold that defines twilight.}
}
\value{
A dataframe with one row for each intervals and columns
\item{\code{Start}}{the start of the interval as POSIXct}
\item{\code{Start}}{the end of the interval as POSIXct}
\item{\code{Midpoint}}{the midpoint of the interval as POSIXct}
}
\description{
Find intervals of darkness spanning more than 24 hours.
}
\details{
The \code{\link{findTwilights}} function find twilight times by
searching for adjacent nights, but is unable to search beyond a
period of light or dark spaaning greater than 24 hours.  This function
finds extended periods of light or dark that span more than 24 hours.
}

