% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{findCrepuscular}
\alias{findCrepuscular}
\title{Search for crepuscular light segments}
\usage{
findCrepuscular(tagdata, twilights, min.threshold = 0, max.threshold = 60,
  extend.dark = 1, extend.light = 2)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{twilights}{a dataframe generated by
\code{\link{findTwilights}}.}

\item{min.threshold}{the light level that defines the dark end of
the crepuscular period.}

\item{max.threshold}{the light level that defines the light end of
the crepuscular period.}

\item{extend.dark}{the number of hours to search towards midnight.}

\item{extend.light}{the number of hours to search towards noon.}
}
\value{
A dataframe with columns.
\item{\code{Twilight}}{times of twilight}
\item{\code{Rise}}{logical indicating sunrise}
\item{\code{Start}}{the start of the crepuscular period}
\item{\code{End}}{the end of the crepuscular period}
where each row corresponds to a single twilight.
}
\description{
Search for crepuscular intervals
}
\details{
Using the twilights calculated by \code{findTwilights},
\code{findCrepuscular} searchs for the intervals of reduced light
intensity around twilight and returns the dataframe constructed by
\code{findTwilights} augmented with the start and end times of
the crepuscular segment. The sugset of data corresponding to these
segments can be extracted with \code{extractCrepuscular}.
}

