% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{imageDraw}
\alias{imageDraw}
\title{Image plot with twilights}
\usage{
imageDraw(
  tagdata = NULL,
  twilights = NULL,
  offset = 0,
  xlim = NULL,
  mark = NULL,
  points.col = NULL,
  ribbon.col = NULL,
  zlim = c(0, 64),
  point.cex = 0.6
)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{twilights}{dataframe of twilight times as generated by
\code{\link{findTwilights}}.}

\item{offset}{the starting hour for the vertical axes.}

\item{xlim}{the range of dates to plot.}

\item{mark}{twilights to mark with a cross}

\item{points.col}{colours of twilight points}

\item{ribbon.col}{colours of twilight ribbons}

\item{zlim}{the range of light levels to plot.}

\item{point.cex}{expansion factor for plot points.}
}
\description{
Image plot with twilight points and ribbons
}
\details{
Draws a light image with twilight points or twilight ribbon. If
tagdata is \code{NULL}, no image is drawn.  Twilight points and
ribbons are only drawn when twilight data is provided and the
point or ribbon colour specifications are non \code{NULL}.
}
