% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{intervalData}
\alias{intervalData}
\alias{twilightData}
\title{Extract auxiliary tag data}
\usage{
intervalData(date, y, start, end, FUN = function(y) if (all(is.na(y))) NA else
  mean(y, na.rm = TRUE))

twilightData(date, y, twl, before = 4, after = 4, FUN = function(y) if
  (all(is.na(y))) NA else mean(y, na.rm = TRUE))
}
\arguments{
\item{date}{the sequence of sample times as POSIXct.}

\item{y}{the sequence of responses.}

\item{start}{the start of the time interval to summarize.}

\item{end}{the end of the time interval to summarize.}

\item{FUN}{summary function}

\item{twl}{the sequence of estimated twilight times as POSIXct.}

\item{before}{the interval (hours) before twilight to include in the interval.}

\item{after}{the interval (hours) after twilight to include in the interval.}
}
\value{
the data summary for each twilight.
}
\description{
Summarise auxiliary data in specified intervals
}
\details{
Given a time series of responses and a sequence of start and end
times defining time intervals spanned by the time series,
\code{intervalData} computes a summary statistic for each
interval. The \code{twilightData} is similar, but the time
intervals are defined as a number of hours before and after a
given sequence of twilight times.
}
