% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAStag.R
\name{lightImage}
\alias{lightImage}
\title{Display Light Data as an Image}
\usage{
lightImage(tagdata, offset = 0, zlim = c(0, 64), dt = NA, xlab = "Date",
  ylab = "Hour", col = grey(seq(0, 1, length = 256)), ...)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{offset}{the offset for the vertical axis in hours.}

\item{zlim}{the range of light levels to plot.}

\item{dt}{the time interval to which the data are resampled (secs).}

\item{xlab}{the x axis label.}

\item{ylab}{the y axis label.}

\item{col}{the colour palette}

\item{...}{additional arguments to pass to image.}
}
\value{
Returns the date and hour coordinates of the image grid.
}
\description{
Display light series as an image.
}
\details{
The \code{lightImage} function displays sequence of light
samples into 24 hour periods, packs these into the columns of a
matrix and displays the result as an image.

This function is essentially a wrapper for \code{tsimage}.
}

