% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactive.R
\name{ndcClosest}
\alias{ndcClosest}
\alias{ndcTsimageDate}
\title{Graphics coordinate conversion}
\usage{
ndcClosest(x, y, xs, ys)

ndcTsimageDate(x, y)
}
\arguments{
\item{x}{ndc x coordinate}

\item{y}{ndc y coordinate}

\item{xs}{points user x coordinates}

\item{ys}{points user y coordinates}
}
\value{
index of closest point
}
\description{
Convert ndc coordinates
}
\details{
Given user coordinates of a sequence of points and the ndc
coordinates (x,y), \code{ndcClosest} returns the index of the
closest point to (x,y).  Given the ndc coordinates of a pixel in a
tsimage plot, ndcTsimageDate returns the date of the pixel.
}

