% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{overlayTwilightResiduals}
\alias{overlayTwilightResiduals}
\title{Overlay data for path editing}
\usage{
overlayTwilightResiduals(twilights, index, mode, xlim, ylim,
  twilight.contours = c(10, 20, 50), zenith = NULL)
}
\arguments{
\item{twilights}{the full twilight data}

\item{index}{the index of the current twilight}

\item{mode}{the overlay mode requested by \code{pathEdit}}

\item{xlim}{the horizontal limits of the map}

\item{ylim}{the vertical limits of the map}

\item{twilight.contours}{the contour levels to display}

\item{zenith}{the solar zenith angle that defines twilight}
}
\description{
Overlay a data when editing an initial path.
}
\details{
The \code{pathEdit} function allows the user to specify a
function to overlay data on the map to assist is selecting an
initial path.

This function overlays a contour map of the difference between the
observed twilight and true time of twilight for that location and
day, where positive and negative contours are shown in different
colours.  This can be useful is selecting an initial path for
which the residuals are positive.
}
\seealso{
\code{\link{pathEdit}}
}

