% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{preprocessLight}
\alias{preprocessLight}
\title{Interactively derive twilight}
\usage{
preprocessLight(tagdata, threshold, offset = 0, lmax = 64, zlim = c(0,
  lmax), extend = 0, dark.min = 0, zenith = 96, fixed = NULL,
  map = FALSE, plotMap = function(xlim, ylim) {     plot.new()    
  plot.window(xlim, ylim) }, twilights = NULL, stage = 1, point.cex = 0.8,
  width = 12, height = 4, palette = defaultPalette[c(5, 2, 9, 3, 4, 1,
  13)])
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)}

\item{threshold}{the light threshold that defines twilight.}

\item{offset}{the starting hour for the vertical axes.}

\item{lmax}{the maximum light level to plot.}

\item{zlim}{the range of light levels to plot in images.}

\item{extend}{a time in minutes. The function seeks periods of
darkness that differ from one another by 24 hours plus or minus
this interval.}

\item{dark.min}{a time in minutes. Periods of darkness shorter
than this interval will be excluded.}

\item{zenith}{the solar zenith angle that defines twilight.}

\item{fixed}{a two column array of fixed locations}

\item{map}{whether to plot an approximate path}

\item{plotMap}{a function to plot a background map when \code{map=TRUE}.}

\item{twilights}{a result of a previous run to re-edit}

\item{stage}{when re-editing, the stage to commence at}

\item{point.cex}{expansion factor for plot points.}

\item{width}{width of the interface windows.}

\item{height}{height of the interface windows.}

\item{palette}{a colour palette of 7 colours.}
}
\value{
A dataframe with columns
\item{\code{Twilight}}{times of twilight}
\item{\code{Rise}}{logical indicating sunrise}
\item{\code{Deleted}}{logical indentifying deleted twilights}
\item{\code{Marker}}{integer vector of marks}
\item{\code{Inserted}}{logical vector of marks}
\item{\code{Twilight3}}{Twilight from stage 3}
\item{\code{Marker3}}{Marker from stage 3}
where each row corresponds to a single twilight.
}
\description{
Interactively derive and edit twilight times.
}
\details{
This function allows the user to interactively search for and edit
twilight times corresponding to a given light threshold. The
process consists of four stages,
\enumerate{
\item Subset -- selection of a subset of the data for processing
\item Search -- semi-automated search for the twilights
\item Insert -- optionally, twilights are inserted where the light record is incomplete
\item Edit -- optioanlly, individual twilights are manually adjusted based on the light profiles.
}

At each stage, user is presented with two windows.  The first
shows the data in its entirety and allows the user to select a
subset of the data to be manipulated in the second window.

In the first stage the user can restrict the data to a subset for
processing.  The first window shows a light image and a coloured
rectangle above the image shows the selected subset of data.  In
the first window, clicking with either mouse button produces a
zoomed image of the clicked region in the second window.  In
either window, left mouse clicks define the start of the selected
subset and right mouse clicks defines the end of the selected
subset. The "a" key accepts the current selection an proceeds to
the next stage.

In either window
\tabular{ll}{
'q' \tab Quits\cr
'a' \tab Accepts any changes and advances to the next stage \cr
'+'/'-' \tab Zoom in or out \cr
}

In the second stage the user guides an semi-automated search for
the twilight times.  Again, a right mouse click in the first
window produces a zoomed image of the clicked region in the
second.  In the second window, the user must click in periods of
nights.  Left mouse clicks select regions to be searched for
twilights, right mouse clicks select regions that should not be
searched.  The "u" key provides an undo facility, allowing mouse
clicks to be deleted, and the "a" key terminates the search and
proceeds to the next stage.

In either window
\tabular{ll}{
'q' \tab Quits, returning the dataframe of edited twilights\cr
'a' \tab Accepts any changes and advances to the next stage \cr
'u' \tab Removes the most recent search point\cr
'+'/'-' \tab Zoom in or out \cr
}

In the third stage the user may insert additional twilights where
the light record is incomplete. As for the first two stages, right
mouse clicks in the first window produce a zoomed image of the
clicked region in the second window.  In the second window, the
user may ad times of sunrise or sunset.  A left mouse click adds a
sunset, and a right mouse click adds a sunrise. The numeric keys
on the keyboard mark the most recent inerted point with a "mark" 0
to 9. If the user provides both the \code{zenith} argument and a
two column matrix of locations for the \code{fixed} argument,
marking a point with the digit n replaces the marked twilight with
the computed twilight time for the n-th fixed location.  Again "u"
provides an undo facility and the "a" key proceeds to the next
stage.

In either window
\tabular{ll}{
'q' \tab Quits, returning the dataframe of edited twilight segments \cr
'a' \tab Accepts any changes and advances to the next stage \cr
'u' \tab Removes the most recent insertion \cr
'+'/'-' \tab Zoom in or out \cr
'0'-'9' \tab Mark the most recent addition \cr
}

In the fourth stage, the user may adjust individual twilights
based on the observed light profile.  The first window now shows
the sequence of twilights.  A right mouse click on any twilight
shows the light profile for that twilight in the second window,
together with the profiles for the preceeding and following days.
A left click in the second window proposes a new location for the
current twilight, but no change is made until the edit is accepted
with the "a" key. A right click in the second window marks the
twilight as deleted, and the numeric keys mark the twilight with a
"mark" 0 to 9. If the user provides both the \code{zenith}
argument and a two column matrix of locations for the \code{fixed}
argument, marking a point with the digit n replaces the marked
twilight with the computed twilight time for the n-th fixed
location.  The "u" key reverts the edits to the current twilight.

In either window
\tabular{ll}{
'q' \tab Quits, returning the dataframe of edited twilight segments \cr
'a' \tab Accepts the candidate edit \cr
'd' \tab Toggle deletion of this twilight \cr
'i' \tab Toggles the display of the light image \cr
'p' \tab Toggles the display of individual points \cr
'p' \tab Toggles the display of surrounding profiles \cr
'r' \tab Resets the selection \cr
'u' \tab Revert changes to this twilight \cr
'+'/'-' \tab Zoom in or out \cr
'Left arrow' \tab Jump to previous twilight \cr
'Right arrow' \tab Jump to ext twilight \cr
'0'-'9' \tab Mark this twilight \cr
}

If \code{map=TRUE} an additional window is constructed that shows
the approximate track based on hte current estimated twilights. A
function to draw a background map for this window can be supplied
with the \code{plotMap} argument.  This must be ia function of two
arguments \code{xlim} and \code{ylim} the determine the extent of
the plot. A left mouse click in this window selects the twilight
associated with the nearest location, a right mouse click in this
window displays the time of twilight at the selected location for
the currently selected date as an edit in the light profile
window, and this edit can be accepted with the "a" key.

The results of a previous call of \code{preprocessLight} can be
updated by providing the dataframe generated by the previous call
as the \code{twilights} argument, and using the \code{stage}
argument to determine to point from which to restart the process.
}
