% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactive.R
\name{profileInit}
\alias{profileInit}
\alias{profileOverlay}
\title{Light profile plot}
\usage{
profileInit(date, light, lmax = 64, xlab = "", main = "")

profileOverlay(date, light, threshold = NULL, point = FALSE, lag = TRUE,
  profile.col = defaultPalette[c(9, 5, 2)],
  threshold.col = defaultPalette[1], point.cex = 0.6)
}
\arguments{
\item{date}{a list of three sequences of sample times as POSIXct.}

\item{light}{a list of three sequences of observed light levels}

\item{lmax}{the maximum light level.}

\item{xlab}{the x axis label.}

\item{main}{the main title.}

\item{threshold}{threshold levels to display}

\item{point}{show individual observations as points?}

\item{lag}{show profiles from neighbouring days}

\item{profile.col}{the colours of the three light profiles.}

\item{threshold.col}{the colour of the threshold markers.}

\item{point.cex}{expansion factor for plot points.}
}
\description{
Initialize a light profile plot for plotting
}
\details{
Sets up axes for drawing light profile plots.
}

