% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAStag.R
\name{readAct}
\alias{readAct}
\alias{readAct2}
\title{Read Activity Data}
\usage{
readAct(file, skip = 0)

readAct2(file, d.lig = NULL, skip = 0)
}
\arguments{
\item{file}{the activity file to import.}

\item{skip}{number of initial lines to skip}

\item{d.lig}{dataframe of light data created with \code{readLig}}
}
\value{
\code{readAct} returns a dataframe with columns
\item{\code{Valid}}{an indicator of whether the record is valid}
\item{\code{Date}}{the date and time of the observation}
\item{\code{Julian}}{date and time as Julian date}
\item{\code{Activity}}{the recorded activity level}
\item{\code{Temp}}{the recorded temperature}
If \code{readAct2} is passed a \code{d.lig} dataframe, it returns that
dataframe with an appended column
\item{\code{Activity}}{the number of seconds the tag has been wet.}
otherwise it returns a dataframe with columns
\item{\code{Valid}}{an indicator of whether the record is valid}
\item{\code{Date}}{the date and time of the observation}
\item{\code{Julian}}{date and time as Julian date}
\item{\code{Activity}}{the number of seconds of activity}
\item{\code{Wet}}{whether the tag was wet or dry}
}
\description{
Read the activity data from a BAS tag
}
\details{
Functions for importing activity data from BAS tags.  The
\code{readAct} function imports activity data from an early
generation BAS tag while \code{readAct2} imports activity data
from later generation BAS tags that run length encode activity.
If \code{readAct2} is passed a dataframe constructed with
\code{readLig}, it interpolates activity to the times at which
light is recorded and returns a merged dataframe with both light
and activity data.
}

