% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAStag.R
\name{readLig}
\alias{readLig}
\title{Read Light Data}
\usage{
readLig(file, skip = 0)
}
\arguments{
\item{file}{the light file to import.}

\item{skip}{number of initial lines to skip}
}
\value{
Returns a dataframe with columns
\item{\code{Valid}}{an indicator of whether the record is valid}
\item{\code{Date}}{the date and time of the observation}
\item{\code{Julian}}{date and time as Julian date}
\item{\code{Light}}{the recorded light level}
}
\description{
Read light data from a BAS tag
}
\details{
The \code{readLig} function imports the light record from a ".lig"
file generated by a BAS tag.
}

