% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{readMTdeg}
\alias{readMTdeg}
\title{Read Temperature Data}
\usage{
readMTdeg(file, skip = 20)
}
\arguments{
\item{file}{the sst file to import.}

\item{skip}{number of initial lines to skip}
}
\value{
Return a dataframe with columns
\item{\code{Date}}{the date and time of the observation}
\item{\code{MinSST}}{the minimum SST}
\item{\code{MaxSST}}{the maximum SST}
\item{\code{SST}}{the mean SST}
\item{\code{Wet}}{the number of wet samples}
\item{\code{Conductivity}}{the condicivity}
}
\description{
Read Temperature data from a Migrate Technologies tag
}
\details{
The \code{readMTdeg} function reads temperature data from a
Migrate Technologies tag.
}
