% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{readMTlux}
\alias{readMTlux}
\title{Read Light Data}
\usage{
readMTlux(file, skip = 20)
}
\arguments{
\item{file}{the lux file to import.}

\item{skip}{number of initial lines to skip}
}
\value{
Returns a dataframe with columns
\item{\code{Date}}{the date and time of the observation}
\item{\code{Light}}{the recorded light level}
}
\description{
Read light data from a Migrate Technologies tag
}
\details{
The \code{readMTlux} function imports the light record from a
".lux" file generated by a Migrate Technologies tag.  These files
have an extensive header, and it may be necessary to set the
\code{skip} argument to skip the lines including the two column
headers.
}
