% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{readMTsst}
\alias{readMTsst}
\title{Read Temperature Data}
\usage{
readMTsst(file, d.lux = NULL, skip = 20)
}
\arguments{
\item{file}{the sst file to import.}

\item{d.lux}{dataframe of light data created with \code{readMTlux}}

\item{skip}{number of initial lines to skip}
}
\value{
If \code{readMTsst} is passed a "lux" dataframe, it returns that
dataframe with the appended columns
\item{\code{MinSST}}{the minimum SST}
\item{\code{MaxSST}}{the maximum SST}
\item{\code{SST}}{the mean SST}
\item{\code{NSST}}{the number of SST samples}
otherwise it returns a dataframe with columns
\item{\code{Date}}{the date and time of the observation}
\item{\code{MinSST}}{the minimum SST}
\item{\code{MaxSST}}{the maximum SST}
\item{\code{SST}}{the mean SST}
\item{\code{NSST}}{the number of SST samples}
}
\description{
Read SST data from a Migrate Technologies tag
}
\details{
The \code{readMTsst} function reads temperature data from a
Migrate Technologies tag.  If \code{readMTsst} is passed a
dataframe constructed with \code{readMTlux}, it matches the
recorded temperatures to the closest recorded light reading and
returns the merged dataframe with both temperature and activity
data.
}
