% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAStag.R
\name{readTem2}
\alias{readTem2}
\title{Read Temperature Data}
\usage{
readTem2(file, d.lig = NULL, skip = 0)
}
\arguments{
\item{file}{the temperature file to import.}

\item{d.lig}{dataframe of light data created with \code{readLig}}

\item{skip}{number of initial lines to skip}
}
\value{
If \code{readTem2} is passed a "lig" dataframe, it returns that
dataframe with the appended column
\item{\code{Temperature}}{the recorded temperature}
otherwise it returns a dataframe with columns
\item{\code{Valid}}{an indicator of whether the record is valid}
\item{\code{Date}}{the date and time of the observation}
\item{\code{Julian}}{date and time as Julian date}
\item{\code{Temperature}}{the recorded temperature.}
}
\description{
Read temperature data from a BAS tag
}
\details{
The \code{readTem2} function reads temperature data from a later
generation BAS tag.  If \code{readTem2} is passed a dataframe
constructed with \code{readLig}, it matches the recorded
temperatures to the closest recorded light reading and returns the
merged dataframe with both temperature and activity data.
}

