% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interactive.R
\name{selectData}
\alias{selectData}
\title{Select data}
\usage{
selectData(date, r, deleted = NULL, extend = 48, xlab = "Date",
  ylab = "", width = 12, height = 4, palette = defaultPalette[c(2, 1)],
  ...)
}
\arguments{
\item{date}{sample times.}

\item{r}{recorded response.}

\item{deleted}{initial logical vector indicating which
observations are to be deleted.}

\item{extend}{the period (in hours) before and after the selection
to be shown in the zoom window.}

\item{xlab}{label for the x axis}

\item{ylab}{label for the y axis}

\item{width}{width of the interface windows.}

\item{height}{height of the interface windows.}

\item{palette}{a colour palette of 4 colours.}

\item{...}{additional parameters to pass to plot.}
}
\value{
a logical vector that indicates the observations to be deleted.
}
\description{
Interactively delete observations from a time series
}
\details{
Interactively delete observations from a time series.  The time
series is displayed. Left mouse button clicks zoom on a region of
the plot; right mouse button clicks toggle deletion of individual
points.

\tabular{ll}{
'q' \tab Quits, returning a indicator vector \cr
'r' \tab Reset zoom to full plot range\cr
's' \tab Toggle reseting of y axes scale on zoom\cr
'+'/'-' \tab Adjust the level of zoom\cr
'Left/Right arrow' \tab Jump forward or backward in hte time series \cr
}
}

