% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{selectionRectangle}
\alias{selectionRectangle}
\title{Selection rectangle}
\usage{
selectionRectangle(x1, x2, col, add = FALSE)
}
\arguments{
\item{x1}{coordinates of the left boundaries of rectangles}

\item{x2}{coordinates of the right boundaries of rectangles}

\item{col}{rectangle colours}

\item{add}{add rectangles to existing rectangles.}
}
\description{
Draw rectangle above plot
}
\details{
Draws a sequence of rectangles above a plot to indicate sections
of the data that have been selected.  Drawing the rectangle above
the plot means the selection can be updated without having to
redraw the main plot.
}

