% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BAStag.R
\name{thresholdCalibrate}
\alias{thresholdCalibrate}
\title{Light Threshold Calibration}
\usage{
thresholdCalibrate(tagdata, lon, lat, max.adjust = TRUE,
  col = c("dodgerblue", "firebrick"), ...)
}
\arguments{
\item{tagdata}{a dataframe with columns \code{Date} and
\code{Light} that are the sequence of sample times (as POSIXct)
and light levels recorded by the tag.}

\item{lon}{calibration longitude.}

\item{lat}{calibration latitude.}

\item{max.adjust}{adjust twilights for tags that report the
maximum light interval observed in the preceeding sampling
interval.}

\item{col}{colors for increasing and decreasing zenith angles}

\item{...}{additional arguments to pass to \code{plot}.}
}
\description{
Calibrate threshold light level to solar zenith angle
}
\details{
Given data for a known location, this plots the recorded light
against the solar zenith angle to allow observed light levels to
be calibrated to solar zenith angle.
}

