% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{tsimage}
\alias{tsimage}
\alias{tsimageLocator}
\title{Display Tag Data as an Image}
\usage{
tsimage(date, y, offset = 0, dt = NA, xlab = "Date", ylab = "Hour",
  xaxt = par("xaxt"), ...)

tsimageLocator(im, n = 512)
}
\arguments{
\item{date}{the sequence of sample times as POSIXct.}

\item{y}{the sequence of responses.}

\item{offset}{the starting hour for the vertical axis.}

\item{dt}{the time interval to which the data are resampled (secs).}

\item{xlab}{the x axis label.}

\item{ylab}{the y axis label.}

\item{xaxt}{a character that specifies the x axis type (see \code{par}).}

\item{...}{additional arguments to be passed to \code{image}.}

\item{im}{the axes coordinates returned by \code{tsimage}.}

\item{n}{the number of points to select.}
}
\value{
\code{tsimage} returns the date and hour coordinates of the image
grid.
\code{tsimageLocator} returns the times corresponding to the
selected pixels.
}
\description{
Display time series data as an image.
}
\details{
The \code{tsimage} function divides a sequence of samples into 24
hour periods, packs these into the columns of a matrix and
displays the result as an image.  The data are resampled to ensure
the time interval between samples is regular and evenly divides
the 24 hour period. The time interval to which the data are
resampled is selected automatically, or can be specified with the
\code{dt} argument.

The \code{tsimageLocator} function allows the user to select
pixels with the mouse pointer in the style of \code{locator}, and
returns the times corresponding to the selected pixels.

These functions ignore clock drift and assume the sampling period
exactly divides 24 hours.
}

