% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAStag.R
\name{tsimagePlot}
\alias{tsimagePlot}
\alias{tsimagePoints}
\alias{tsimageLines}
\alias{tsimageRibbon}
\title{tsimage plot}
\usage{
tsimagePlot(date, offset, xlab = "Date", ylab = "Hour", ...)

tsimagePoints(date, offset, ...)

tsimageLines(date, offset, ...)

tsimageRibbon(date1, date2, offset, ...)
}
\arguments{
\item{date}{times as a vector of POSIXct.}

\item{offset}{the starting hour for the vertical axes.}

\item{xlab}{the label for the x axis.}

\item{ylab}{the label for the y axis.}

\item{...}{additional arguments parameters to pass to \code{plot},
\code{points} or \code{lines}.}

\item{date1}{lower bound as a vector of POSIXct.}

\item{date2}{upper bound as a vector of POSIXct.}
}
\description{
Plot times as hour vs date
}
\details{
Analogs of \code{plot}, \code{points} and \code{lines} that plot
the time of day against date.  The \code{tsimageRibbon} uses
\code{polygon} to produce a polygonal "ribbon" that represents a
series of daily time intervals.
}
